+++ planet-mitchell/src/Function.hs
@@ -9,6 +9,10 @@ module Function
   , loop
   , on
   , until
+  , curry
+  , curry3
+  , uncurry
+  , uncurry3
   , Endo(..)
   , Op(..)
   ) where
@@ -17,4 +21,6 @@ import Control.Monad.Extra        (loop)
 import Data.Function
 import Data.Functor.Contravariant
 import Data.Semigroup             (Endo(Endo, appEndo))
+import Data.Tuple                 (curry, uncurry)
+import Data.Tuple.Extra           (curry3, uncurry3)
 import GHC.Base