+++ planet-mitchell/src/IO.hs
@@ -6,6 +6,10 @@ module IO
   , timeout
   , once
   , onceFork
+    -- * Memoized IO
+  , Memoized
+  , runMemoized
+  , memoizeMVar
     -- * IO exceptions
   , module System.IO.Error
   ) where
@@ -14,4 +18,5 @@ import Control.Concurrent.Extra (once, onceFork)
 import Control.Monad.IO.Class   (MonadIO(..))
 import System.IO
 import System.IO.Error
+import UnliftIO.Memoize         (Memoized, memoizeMVar, runMemoized)
 import UnliftIO.Timeout         (timeout)